+++ example/lib/data_model.dart
@@ -1,7 +1,0 @@
-class Data {
-  final int id;
-  final String title;
-  final String subTitle;
-
-  Data(this.id, this.title, this.subTitle);
-}